file_previews - some restructuring
